Moved each model into own python file.
